Address CI review on test coverage closures (#545)
Three blockers from the previous CI review, all fixed:

- TEST-1: moved `shouldDropPersistedTab` out of `CodeBrowserView.tsx`
  into a dedicated `apps/web/src/lib/persisted-tab-self-heal.ts`. The
  helper was exported from the component solely so the unit test
  could import it — TEST-1 explicitly bans that shape. It now lives
  alongside its sibling `dispatch-open-file.ts` in `lib/`, with the
  test importing from the new module.
- TEST-21: added `WorkspacePage.waitForMobileReady()` encapsulating
  the `getByRole("button", { name: "Files" })` locator the mobile
  spec previously called directly from the test body. Both mobile
  tests now go through the page object.
- TEST-24: rewrote the mobile spec's negative assertion. Instead of
  `waitForTimeout(800) + isVisible().toBe(false)`, the test now
  dispatches the cross-workspace event and runs
  `expect.poll(() => dialog.isVisible(), { timeout: 800 }).toBe(true)`
  inside a try/catch, then asserts the poll DID time out without
  observing a flip to visible. This is positive-shaped per TEST-23
  (auto-retrying poll for `true`), and the bounded timeout makes
  the negative deterministic: the bug version's autoOpen lands
  within ~200 ms and the assertion fails fast, while the fix
  version's listener drops the event and the poll exhausts its
  budget. The positive control above establishes that the
  listener's react/render/search lifecycle is well under our
  budget, so a regression that DID accept the cross-workspace
  event would land its visibility flip inside the 800 ms window.

Plus the nit: dropped the TEST-24 rule-ID citation from the
explanatory comment (CODE-19) in favour of inline rationale.
